requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-richtextfield==1.6.1
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
psycopg2-binary==2.8.6
